format icons using flex
used margin instead of flex between h2 and the icons themselves because felt unnecessary. I couldn't find a fontawesome icon for gatsby js (wanted to add that one in too) but we'll leave for now. as noted in the to-do comment, want to generate icons programmatically in js because it was very annoying to type out