redesign portfolio (#4)
* delete all unused icons and reset index.html

* link css, add header divs with links

although the 'HL' doesn't really feel like it's needed right now, as its purpose is to link to this homepage, i think it's possible that in the future i'll have more pages in my portfolio, so this is helpful

* create css file and reset default styles

* update background colour of entire body

* remove anchor (link) styling and update colour of text to white

* make header a flex container and space out text

* add padding

* fix: make header responsive

by using hardcoded padding, header elements are crushed and misaligned on smaller, thinner screens. instead, by using width 50vw (50% of viewport's width) this works on smaller screens (to a certain degree)

* import fonts from Google Fonts, and update padding

because the font-size is a bit bigger, it is better to have a bigger padding-top. Updated colour from white to a more grey colour. 'Tenor Sans' only comes in one font-weight - any thicker/bolder actually looks quite ugly and pixellated so I did not style it

* add section

* add copy of hero, and make hero as wide as header

* update colours and add class to self-introduction

helps distinguish from the other <p> inside <hgroup> for selector

* make subheading a <span> within <h1>

this makes it distinguishable still, however, does not cause misalignment in styles by applying, for example, font-size on the hgroup. this helps simplify the css codebase

* update colour of subheading to darkgrey, and enforce linebreaks

used display: inline-block instead of hardcoding <br> for consistency. removed class name that was unused in css

* remove unused class name and update css

* reduce fontsize of header group

* enforce newline in hero text

* increase line height of self-introduction paragraph

* add margin and simplify copy

increase space between hero heading group and the paragraph

* update colours of text

align colours with the monochromatic theme

* update colour scheme to black bird from canva

* tweak colours for readability and consistency

* update margin to increase space between header and hero

* reduce fontsize of header

* update line height for readability

* reduce width of hero section

* update margin-top of the paragraph to align text heights

* add margin to each section and add h2 and styling

* reduce width of total page and align between page and hero

previously, hero was narrower

* update colours and add article (semantic html) of list of skills

* add icons

didn't make these animated like before because wanted a more minimalist look

* format icons using flex

used margin instead of flex between h2 and the icons themselves because felt unnecessary. I couldn't find a fontawesome icon for gatsby js (wanted to add that one in too) but we'll leave for now. as noted in the to-do comment, want to generate icons programmatically in js because it was very annoying to type out

* increase margin between h2 and icons

* delete icons as unused

* add contact section and heading

* add contact methods

* make contact section narrower

however, this is not right, as the text should be all left-aligned

* make the right column of table 3x as big as left columnn

* wrap when screen is very narrow

* on narrower screens, skills icons collapse

this is done by using flex-wrap, with a gap (though this feels really hacky - the icons look misaligned)

* add links in contact via anchors

no need for target=_blank for email because this doesn't usually open in a browser window, but in a mail application (desktop) of choice.

* use max-width instead of vw units for body

this avoids having everything resize as a result of screen narrowing. added padding for better look on smaller screens

* reduce space between bottom of header and hero

* add colour change when hovered over links in contact

used a very subtle colour change, and also applied to the header .right 'contact' button that takes user to the contact section

* update hover colour on links

the previous colour had shades that made it look a bit odd. now, i just used generic whitesmoke. i did not use white because it would be a bit too harshly contrasted with the current colour

* update copy

* restructure html

add class to the subheading. the class to the name is because i'm about to style that differently (maybe in cursive) to bring the portfolio to life a bit more

* import sacramento google font

this provides a 'cursive' font for us to use

* style just the name

* Revert "style just the name"

This reverts commit f0c23a91e86b5251cd9e057198257049eac83336.

* update colours

* add flex table and rows from contact

this ensures the width are aligned, as they both picked up by selectors on class in style.css

* add copy for cba

seems like i can't put an unordered list inside of a p tag so i attached the .right class to the ul tag directly

* update colours and fonts of cba entry, darkened colour on the left side of the .table (including in the .contact section) to better distinguish .left from .right

* restore hover effect, update colour of description of job

* add telstra copy and adjust gap

* fix: ensure that contact section's table was being selected - used class selector for contact to align with other sections

* update line-height for description of job for legibility

* enforce linebreaks

this avoids scenarios where there's only one word on a single line

* update spacing and font size

* add acc entry

* add stats about speeding up scripts

* update copy and enforce linebreaks

* update colour of subheading in hero

* add section

* remove unncessary closing div tag

* update contact as id because we need to link to it from header

* change colours

* use small semantic html for skills section

* update font for skills section

* add favicon from calculator project

* add background colour to calculator project

* add comments to css to separate selectors etc

* create more flex containers that are also flex items

easier to control gaps between items using 'gap' than margins i think

* reduce size of left side icon

* add icons, rename screenshot and favicon for calculator

having a bigger image might be more engaging than just a favicon that might not be thought of as having any relation with the actual project

* removed text - just used icons

* update alt text and fix broken links

did not have https:// prefix, causing browser to think that it is a subpage of the current site

* fix: text no horizontally aligned between left side and right side of experiences section

this is potentially due to the different fonts. there's no margin-top nor padding-top on the .left to cause it to be pushed down slightly

* resize images, and add huaxia as a project

* updating projects copy

exclamation marks in excess looks cheap

* update gaps

* reduce size of images and update capitalisation in copy for Huaxia

* add bjj-countdown as project, including screenshot

* fix: alt text

* add stoic quotes screenshot and copy

* update screenshot of stoic quotes to align with rest of the screenshots (16:9 ratio)

* add favicon

* reorder projects

removed bjj countdown for now - perhaps will add back once i enhance it some more. i want to keep stoic quotes because i implemented a stack data structure in it to ensure the same quote isn't shown for at least 5 quotes. i want to talk about this in interviews if possible