update copy and style
added, at the top of my experiences, being a web developer for Huaxia. This is primarily what I'm doing right now while being on extended leave from CBA. It is also the most relevant for my portfolio for being a web developer. I also added that I'm a web designer, because I do love that a lot. I then updated the line-height of the paragraphs because it makes it easier to read.