swap job position and employer (#17)
* update html by swapping elements around; got rid of ACC job

to be fair, I was very early on in my web design journey for the ACC gig. it's better I just leave off for now

* update copy and styling

* add link to huaxia's website in experience section, and use not selector to simplify link selection

we want all links (anchors) that isn't an image in the projects to have the underline animation). detected bug that the gradient on the images are not showing

* make link to huaxia site bold

* fix: not selector for anchors fixed