update email and realign flexbox table
previously applied a margin-top property to all text in the right side of the flexbox 'table' in experience and contact to fix just the experience section, this ended up pushing the 'contact' right-hand-side text down due to the margin-top. applied specifically only to the experience section and is now fixed.